Move util function; Add tests #83
