+++ src/main/js/Navigation.js
@@ -72,12 +72,4 @@ class Navigation extends React.Component<Props, State> {
 
 }
 
-export function pathWithTrailingSlash(path) {
-    var lastChar = path.substr(-1);
-    if (lastChar !== '/') {
-        path = path + '/';
-    }
-    return path;
-}
-
 export default Navigation;
