+++ src/main/js/repository/components/Branch.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import {Link} from 'react-router-dom';
 import { withRouter } from "react-router";
-import {pathWithTrailingSlash} from "../../Navigation";
+import {pathWithTrailingSlash} from "../../util";
 
 type Props = {
   branch: any
