+++ src/main/js/util.js
@@ -1,0 +1,8 @@
+
+export function pathWithTrailingSlash(path) {
+    var lastChar = path.substr(-1);
+    if (lastChar !== '/') {
+        path = path + '/';
+    }
+    return path;
+}
