+++ src/main/js/util.test.js
@@ -1,0 +1,15 @@
+import {pathWithTrailingSlash} from './util';
+
+test('test pathWithTrailingSlash for path with slash', () => {
+    let path = "foo/";
+    let newPath = pathWithTrailingSlash(path);
+
+    expect(newPath).toBe(path);
+});
+
+test('test pathWithTrailingSlash for path without slash', () => {
+    let path = "foo/baa";
+    let newPath = pathWithTrailingSlash(path);
+
+    expect(newPath).toBe(path + "/");
+});
